scrapper-app/requirements.txt
flask
flask-cors
google-api-python-client
python-dotenv